Moved a lot of things around, generally cleaned up the code.

+++ src/romarth.h
@@ -1,7 +1,9 @@
 #ifndef ROMARTH_H
 #define ROMARTH_H
 
-char *roman_add(char *sum, int sum_size, char *summand1, char *summand2);
-char *roman_subtract(char *differene, int difference_size, char *minuend, char *suptrahend);
+#include <stddef.h> /* For size_t */
+
+char *roman_add(char *sum, size_t sum_size, char *summand1, char *summand2);
+char *roman_subtract(char *differene, size_t difference_size, char *minuend, char *suptrahend);
 
 #endif /* ROMARTH_H */
